move tests to tests.rs
Also combines the two implementation blocks.
