+++ vortex-array/src/arrays/list/mod.rs
@@ -70,7 +70,7 @@ impl VTable for ListVTable {
 ///
 /// - Offsets must be non-nullable integers (i32, i64, etc.)
 /// - Offsets array has length `n+1` where `n` is the number of lists
-/// - List `i` contains elements from `elements[offsets[i]..offsets[i+1]]`  
+/// - List `i` contains elements from `elements[offsets[i]..offsets[i+1]]`
 /// - Offsets must be monotonically increasing
 ///
 /// # Examples
@@ -115,81 +115,6 @@ pub struct ListEncoding;
 pub trait OffsetPType: NativePType + PrimInt + AsPrimitive<usize> + Into<Scalar> {}
 
 impl<T> OffsetPType for T where T: NativePType + PrimInt + AsPrimitive<usize> + Into<Scalar> {}
-
-// A list is valid if the:
-// - offsets start at a value in elements
-// - offsets are sorted
-// - the final offset points to an element in the elements list, pointing to zero
-//   if elements are empty.
-// - final_offset >= start_offset
-// - The size of the validity is the size-1 of the offset array
-
-impl ListArray {
-    fn validate(
-        elements: &dyn Array,
-        offsets: &dyn Array,
-        validity: &Validity,
-    ) -> VortexResult<()> {
-        // Offsets must be of integer type, and cannot go lower than 0.
-        vortex_ensure!(
-            offsets.dtype().is_int() && !offsets.dtype().is_nullable(),
-            "offsets have invalid type {}",
-            offsets.dtype()
-        );
-
-        // We can safely unwrap the DType as primitive now
-        let offsets_ptype = offsets.dtype().as_ptype();
-
-        // Offsets must be sorted (but not strictly sorted, zero-length lists are allowed)
-        if let Some(is_sorted) = offsets.statistics().compute_is_sorted() {
-            vortex_ensure!(is_sorted, "offsets must be sorted");
-        } else {
-            vortex_bail!("offsets must report is_sorted statistic");
-        }
-
-        // Validate that offsets min is non-negative, and max does not exceed the length of
-        // the elements array.
-        if let Some(min_max) = min_max(offsets)? {
-            match_each_integer_ptype!(offsets_ptype, |P| {
-                let max_offset = <P as NumCast>::from(elements.len()).unwrap_or(P::MAX);
-
-                #[allow(clippy::absurd_extreme_comparisons, unused_comparisons)]
-                {
-                    if let Some(min) = min_max.min.as_primitive().as_::<P>() {
-                        vortex_ensure!(
-                            min >= 0 && min <= max_offset,
-                            "offsets minimum {min} outside valid range [0, {max_offset}]"
-                        );
-                    }
-
-                    if let Some(max) = min_max.max.as_primitive().as_::<P>() {
-                        vortex_ensure!(
-                            max >= 0 && max <= max_offset,
-                            "offsets maximum {max} outside valid range [0, {max_offset}]"
-                        )
-                    }
-                }
-            })
-        } else {
-            // TODO(aduffy): fallback to slower validation pathway?
-            vortex_bail!(
-                "offsets array with encoding {} must support min_max compute function",
-                offsets.encoding_id()
-            );
-        };
-
-        // If a validity array is present, it must be the same length as the ListArray
-        if let Some(validity_len) = validity.maybe_len() {
-            vortex_ensure!(
-                validity_len == offsets.len() - 1,
-                "validity with size {validity_len} does not match array size {}",
-                offsets.len() - 1
-            );
-        }
-
-        Ok(())
-    }
-}
 
 impl ListArray {
     pub fn new(elements: ArrayRef, offsets: ArrayRef, validity: Validity) -> Self {
@@ -279,6 +204,78 @@ impl ListArray {
         let adjusted_offsets = sub_scalar(offsets, first_offset)?;
 
         Self::try_new(elements, adjusted_offsets, self.validity.clone())
+    }
+
+    /// A list is valid if the:
+    /// - offsets start at a value in elements
+    /// - offsets are sorted
+    /// - the final offset points to an element in the elements list, pointing to zero
+    ///   if elements are empty.
+    /// - final_offset >= start_offset
+    /// - The size of the validity is the size-1 of the offset array
+    fn validate(
+        elements: &dyn Array,
+        offsets: &dyn Array,
+        validity: &Validity,
+    ) -> VortexResult<()> {
+        // Offsets must be of integer type, and cannot go lower than 0.
+        vortex_ensure!(
+            offsets.dtype().is_int() && !offsets.dtype().is_nullable(),
+            "offsets have invalid type {}",
+            offsets.dtype()
+        );
+
+        // We can safely unwrap the DType as primitive now
+        let offsets_ptype = offsets.dtype().as_ptype();
+
+        // Offsets must be sorted (but not strictly sorted, zero-length lists are allowed)
+        if let Some(is_sorted) = offsets.statistics().compute_is_sorted() {
+            vortex_ensure!(is_sorted, "offsets must be sorted");
+        } else {
+            vortex_bail!("offsets must report is_sorted statistic");
+        }
+
+        // Validate that offsets min is non-negative, and max does not exceed the length of
+        // the elements array.
+        if let Some(min_max) = min_max(offsets)? {
+            match_each_integer_ptype!(offsets_ptype, |P| {
+                let max_offset = <P as NumCast>::from(elements.len()).unwrap_or(P::MAX);
+
+                #[allow(clippy::absurd_extreme_comparisons, unused_comparisons)]
+                {
+                    if let Some(min) = min_max.min.as_primitive().as_::<P>() {
+                        vortex_ensure!(
+                            min >= 0 && min <= max_offset,
+                            "offsets minimum {min} outside valid range [0, {max_offset}]"
+                        );
+                    }
+
+                    if let Some(max) = min_max.max.as_primitive().as_::<P>() {
+                        vortex_ensure!(
+                            max >= 0 && max <= max_offset,
+                            "offsets maximum {max} outside valid range [0, {max_offset}]"
+                        )
+                    }
+                }
+            })
+        } else {
+            // TODO(aduffy): fallback to slower validation pathway?
+            vortex_bail!(
+                "offsets array with encoding {} must support min_max compute function",
+                offsets.encoding_id()
+            );
+        };
+
+        // If a validity array is present, it must be the same length as the ListArray
+        if let Some(validity_len) = validity.maybe_len() {
+            vortex_ensure!(
+                validity_len == offsets.len() - 1,
+                "validity with size {validity_len} does not match array size {}",
+                offsets.len() - 1
+            );
+        }
+
+        Ok(())
     }
 }
 
@@ -393,161 +390,4 @@ impl ListArray {
 }
 
 #[cfg(test)]
-mod test {
-    use std::sync::Arc;
-
-    use arrow_buffer::BooleanBuffer;
-    use vortex_dtype::Nullability;
-    use vortex_dtype::PType::I32;
-    use vortex_error::VortexUnwrap;
-    use vortex_mask::Mask;
-    use vortex_scalar::Scalar;
-
-    use crate::arrays::list::ListArray;
-    use crate::arrays::{ListVTable, PrimitiveArray};
-    use crate::builders::{ArrayBuilder, ListBuilder};
-    use crate::compute::filter;
-    use crate::validity::Validity;
-    use crate::{Array, IntoArray};
-
-    #[test]
-    fn test_empty_list_array() {
-        let elements = PrimitiveArray::empty::<u32>(Nullability::NonNullable);
-        let offsets = PrimitiveArray::from_iter([0]);
-        let validity = Validity::AllValid;
-
-        let list =
-            ListArray::try_new(elements.into_array(), offsets.into_array(), validity).unwrap();
-
-        assert_eq!(0, list.len());
-    }
-
-    #[test]
-    fn test_simple_list_array() {
-        let elements = PrimitiveArray::from_iter([1i32, 2, 3, 4, 5]);
-        let offsets = PrimitiveArray::from_iter([0, 2, 4, 5]);
-        let validity = Validity::AllValid;
-
-        let list =
-            ListArray::try_new(elements.into_array(), offsets.into_array(), validity).unwrap();
-
-        assert_eq!(
-            Scalar::list(
-                Arc::new(I32.into()),
-                vec![1.into(), 2.into()],
-                Nullability::Nullable
-            ),
-            list.scalar_at(0)
-        );
-        assert_eq!(
-            Scalar::list(
-                Arc::new(I32.into()),
-                vec![3.into(), 4.into()],
-                Nullability::Nullable
-            ),
-            list.scalar_at(1)
-        );
-        assert_eq!(
-            Scalar::list(Arc::new(I32.into()), vec![5.into()], Nullability::Nullable),
-            list.scalar_at(2)
-        );
-    }
-
-    #[test]
-    fn test_simple_list_array_from_iter() {
-        let elements = PrimitiveArray::from_iter([1i32, 2, 3]);
-        let offsets = PrimitiveArray::from_iter([0, 2, 3]);
-        let validity = Validity::NonNullable;
-
-        let list =
-            ListArray::try_new(elements.into_array(), offsets.into_array(), validity).unwrap();
-
-        let list_from_iter =
-            ListArray::from_iter_slow::<u32, _>(vec![vec![1i32, 2], vec![3]], Arc::new(I32.into()))
-                .unwrap();
-
-        assert_eq!(list.len(), list_from_iter.len());
-        assert_eq!(list.scalar_at(0), list_from_iter.scalar_at(0));
-        assert_eq!(list.scalar_at(1), list_from_iter.scalar_at(1));
-    }
-
-    #[test]
-    fn test_simple_list_filter() {
-        let elements = PrimitiveArray::from_option_iter([None, Some(2), Some(3), Some(4), Some(5)]);
-        let offsets = PrimitiveArray::from_iter([0, 2, 4, 5]);
-        let validity = Validity::AllValid;
-
-        let list = ListArray::try_new(elements.into_array(), offsets.into_array(), validity)
-            .unwrap()
-            .into_array();
-
-        let filtered = filter(
-            &list,
-            &Mask::from(BooleanBuffer::from(vec![false, true, true])),
-        );
-
-        assert!(filtered.is_ok())
-    }
-
-    #[test]
-    fn test_offset_to_0() {
-        let mut builder =
-            ListBuilder::<u32>::with_capacity(Arc::new(I32.into()), Nullability::NonNullable, 5);
-        builder
-            .append_value(
-                Scalar::list(
-                    Arc::new(I32.into()),
-                    vec![1.into(), 2.into(), 3.into()],
-                    Nullability::NonNullable,
-                )
-                .as_list(),
-            )
-            .vortex_unwrap();
-        builder
-            .append_value(
-                Scalar::list(
-                    Arc::new(I32.into()),
-                    vec![4.into(), 5.into(), 6.into()],
-                    Nullability::NonNullable,
-                )
-                .as_list(),
-            )
-            .vortex_unwrap();
-        builder
-            .append_value(
-                Scalar::list(
-                    Arc::new(I32.into()),
-                    vec![7.into(), 8.into(), 9.into()],
-                    Nullability::NonNullable,
-                )
-                .as_list(),
-            )
-            .vortex_unwrap();
-        builder
-            .append_value(
-                Scalar::list(
-                    Arc::new(I32.into()),
-                    vec![10.into(), 11.into(), 12.into()],
-                    Nullability::NonNullable,
-                )
-                .as_list(),
-            )
-            .vortex_unwrap();
-        builder
-            .append_value(
-                Scalar::list(
-                    Arc::new(I32.into()),
-                    vec![13.into(), 14.into(), 15.into()],
-                    Nullability::NonNullable,
-                )
-                .as_list(),
-            )
-            .vortex_unwrap();
-        let list = builder.finish().slice(2, 4);
-        let list = list.as_::<ListVTable>().reset_offsets().unwrap();
-        assert_eq!(list.len(), 2);
-        assert_eq!(list.offsets().len(), 3);
-        assert_eq!(list.elements().len(), 6);
-        assert_eq!(list.offsets().scalar_at(0), 0u32.into());
-    }
-}
+mod tests;
